[refactor] 디렉토리 구조 리팩토링
## 목적
 가독성과 목적에 맞는 클래스들끼리 구분하기 위해 디렉토리를 변경한다.

## 반영
domain 디렉토리 : 각 기능을 담당하는 클래스들이 담긴다.
finalstring 디렉토리 : enum 클래스가 담긴다.
