+++ src/main/java/baseball/Application.java
@@ -1,6 +1,6 @@
 package baseball;
 
-import baseball.game.Game;
+import baseball.domain.game.Game;
 
 public class Application {
     public static void main(String[] args) {
