+++ src/main/java/baseball/domain/pitch/Pitch.java
@@ -1,4 +1,4 @@
-package baseball.pitch;
+package baseball.domain.pitch;
 
 public class Pitch {
     private final int index;
