+++ src/test/java/baseball/domain/computer/ComputerTest.java
@@ -1,4 +1,4 @@
-package baseball.computer;
+package baseball.domain.computer;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
