+++ src/test/java/baseball/domain/game/GameTest.java
@@ -1,6 +1,6 @@
-package baseball.game;
+package baseball.domain.game;
 
-import baseball.pitch.Pitch;
+import baseball.domain.pitch.Pitch;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
