+++ src/test/java/baseball/domain/hint/HintTest.java
@@ -1,7 +1,7 @@
-package baseball.hint;
+package baseball.domain.hint;
 
-import baseball.game.Game;
-import baseball.pitch.Pitch;
+import baseball.domain.game.Game;
+import baseball.domain.pitch.Pitch;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
