+++ src/test/java/baseball/domain/pitch/PitchTest.java
@@ -1,10 +1,9 @@
-package baseball.pitch;
+package baseball.domain.pitch;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.as;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class PitchTest {
