+++ src/test/java/baseball/domain/valid/ValidImpl.java
@@ -1,4 +1,4 @@
-package baseball.valid;
+package baseball.domain.valid;
 
 public class ValidImpl implements Valid{
 }
